+++ packages/upload-api/test/handlers/store.js
@@ -3,7 +3,6 @@ import * as API from '../../src/types.js'
 import { CID } from 'multiformats'
 import * as CAR from '@ucanto/transport/car'
 import { base64pad } from 'multiformats/bases/base64'
-import * as Link from '@ucanto/core/link'
 import * as StoreCapabilities from '@web3-storage/capabilities/store'
 import { alice, bob, createSpace, registerSpace } from '../util.js'
 import { Absentee } from '@ucanto/principal'
@@ -318,8 +317,6 @@ export const test = {
       }
     )
 
-    assert.deepEqual(Link.parse(item.invocation.toString()), item.invocation)
-
     assert.equal(
       Date.now() - new Date(item.insertedAt).getTime() < 60_000,
       true
@@ -404,10 +401,7 @@ export const test = {
     )
   },
 
-  'store/remove does not fail for non existent link': async (
-    assert,
-    context
-  ) => {
+  'store/remove fails for non existent link': async (assert, context) => {
     const { proof, spaceDid } = await registerSpace(alice, context)
     const connection = connect({
       id: context.id,
@@ -427,21 +421,7 @@ export const test = {
       })
       .execute(connection)
 
-    // expect no response for a remove
-    assert.deepEqual(storeRemove.out.ok, {})
-
-    const storeRemove2 = await StoreCapabilities.remove
-      .invoke({
-        issuer: alice,
-        audience: connection.id,
-        with: spaceDid,
-        nb: { link },
-        proofs: [proof],
-      })
-      .execute(connection)
-
-    // expect no response for a remove
-    assert.deepEqual(storeRemove2.out.ok, {})
+    assert.equal(storeRemove.out.error?.name, 'StoreItemNotFound')
   },
 
   'store/list does not fail for empty list': async (assert, context) => {