+++ packages/upload-api/test/helpers/context.js
@@ -7,7 +7,9 @@ import { DudewhereBucket } from '../storage/dude-where-bucket.js'
 import { ProvisionsStorage } from '../storage/provisions-storage.js'
 import { DelegationsStorage } from '../storage/delegations-storage.js'
 import { RateLimitsStorage } from '../storage/rate-limits-storage.js'
+import { RevocationsStorage } from '../storage/revocations-storage.js'
 import * as Email from '../../src/utils/email.js'
+import { create as createRevocationChecker } from '../../src/utils/revocation.js'
 import { createServer, connect } from '../../src/lib.js'
 import * as Types from '../../src/types.js'
 import * as TestTypes from '../types.js'
@@ -22,6 +24,7 @@ export const createContext = async (options = {}) => {
   const uploadTable = new UploadTable()
   const carStoreBucket = await CarStoreBucket.activate()
   const dudewhereBucket = new DudewhereBucket()
+  const revocationsStorage = new RevocationsStorage()
   const signer = await Signer.generate()
   const id = signer.withDID('did:web:test.web3.storage')
 
@@ -34,6 +37,7 @@ export const createContext = async (options = {}) => {
     provisionsStorage: new ProvisionsStorage(options.providers),
     delegationsStorage: new DelegationsStorage(),
     rateLimitsStorage: new RateLimitsStorage(),
+    revocationsStorage,
     errorReporter: {
       catch(error) {
         assert.fail(error)
@@ -44,6 +48,7 @@ export const createContext = async (options = {}) => {
     uploadTable,
     carStoreBucket,
     dudewhereBucket,
+    ...createRevocationChecker({ revocationsStorage }),
   }
 
   const connection = connect({