+++ packages/upload-api/test/lib.js
@@ -11,6 +11,7 @@ import * as Upload from './handlers/upload.js'
 import { test as delegationsStorageTests } from './storage/delegations-storage-tests.js'
 import { test as provisionsStorageTests } from './storage/provisions-storage-tests.js'
 import { test as rateLimitsStorageTests } from './storage/rate-limits-storage-tests.js'
+import { test as revocationsStorageTests } from './storage/revocations-storage-tests.js'
 import { DebugEmail } from '../src/utils/email.js'
 
 export * from './util.js'
@@ -24,6 +25,7 @@ export const storageTests = {
   ...delegationsStorageTests,
   ...provisionsStorageTests,
   ...rateLimitsStorageTests,
+  ...revocationsStorageTests,
 }
 
 export const handlerTests = {
@@ -45,5 +47,6 @@ export {
   delegationsStorageTests,
   provisionsStorageTests,
   rateLimitsStorageTests,
+  revocationsStorageTests,
   DebugEmail,
 }